app/pyproject.toml
[project]
name = "mypy-playground"
version = "0.0.0"
requires-python = ">=3.14"
dependencies = [
    "aiodocker>=0.13.0,<1.0",
    "fastapi>=0.115.0,<0.116.0",
    "fastapi-cli[standard-no-fastapi-cloud-cli]",
    "google-auth",
    "httpx>=0.28.0,<0.29.0",
    "prometheus-client<1",
    "pydantic>=2.9.0,<3.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "requests",
]

[dependency-groups]
dev = [
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-asyncio>=0.24.0",
    "pytest-cov",
    "pytest-mock",
    "pytest-randomly",
    "types-chardet",
    "types-mock",
    "types-setuptools",
    "types-toml",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = [
    "--cov=mypy_playground",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.14"
strict = true

[[tool.mypy.overrides]]
module = [
    "google.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py314"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # isort
    "I",
    # pyupgrade
    "UP",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-logging
    "LOG",
    # flake8-logging-format
    "G",
    # flake8-simplify
    "SIM",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    # flake8-bandit: assert
    "S101",
]
